requirements.txt
PyGithub>=1.34 # Compatible with 3.6 after 1.34
GitPython
requests
nose
mistune
pyyaml